Recast the Delta to __int64

+++ src/Slaves.cpp
@@ -222,7 +222,7 @@ void GetSlavesXMLSnapShot(const char **ppXMLSnapShot)
          XMLSnapShot += IntToHEXStr<int>(it.first);
          XMLSnapShot += "\"";
          XMLSnapShot += " DELTA=\"";
-         XMLSnapShot += std::to_string(it.second.GetExpireTS() - GetMillis() + it.second.GetTS());
+         XMLSnapShot += std::to_string(__int64(it.second.GetExpireTS() - GetMillis() + it.second.GetTS()));
          XMLSnapShot += "\">";
 		 // Log the relays status as children and not as attributes
 		 for (int i=0; i<NUM_RELAYS; ++i)
